requirements.txt
Django==1.9.1
MySQL-python==1.2.5
Pillow==3.1.1
pytz==2015.7
coreapi==2.3.0
djangorestframework==3.6.2